+++ アクションゲーム/EffectBase.cpp
@@ -34,17 +34,17 @@ void EffectBase::BaseInit(int _maxlife,
 	//座標変化率の設定
 	int division_flame = m_maxlife - m_poschange_frame;
 	if (division_flame <= 0) division_flame = 1;
-	pos_changerate = _pos_amount / (division_flame);
+	pos_changerate = _pos_amount / (float)(division_flame);
 	
 	//回転変化率の設定
 	division_flame = m_maxlife - m_rotchange_frame;
 	if (division_flame <= 0) division_flame = 1;
-	rot_changerate = _rot_amount / (division_flame);
+	rot_changerate = _rot_amount / (float)(division_flame);
 	
 	//スケール変化率の設定
 	division_flame = m_maxlife - m_scalechange_frame;
 	if (division_flame <= 0) division_flame = 1;
-	scale_changerate = _scale_amount / (division_flame);
+	scale_changerate = _scale_amount / (float)(division_flame);
 	
 	endless = _endless;
 	m_live = true;
